Test suite fix (#70)
* Integrity constraint violation SQLSTATE is 40002

This was fixed on Jun2020.

* Fix ambiguous query

On MonetDB Jun2020 the following query fails to parse:

SELECT * FROM table, table

We need to use aliases like:

SELECT * FROM table as t1, table as t2

* Add back the old error code

We would like to support older versions of MonetDB as well. This can
be removed once support for Nov2019 and earlier versions is dropped.