Improve timezone handling (#90)
* Pass client timezone to the server

* Store incoming timezone in datetime instead of shifting time

The TIMESTAMPTZ's the server sends us have already been adjusted to our
timezone.  Any timezone indication we receive should be stored in the
.tzinfo field of the datetime rather than being added to the .hours.

This change makes query `SELECT now()` return the correct datetime
object.  This wasn't the case before.

* Send TIMESTAMPTZ instead of TIMESTAMP when timezone is known

This makes `SELECT %(t)s` work, if `t` is a timestamp with
timezone.  Without this change, we would receive the UTC version
of `t`, without timezone information.

* Add some timezone tests

* Let test_capabilities.py use Tz types where appropriate

* Calling .time on a datetime turns out to lose the timezone!

* Make TIMESTAMPTZ test independent of host timezone

* Move determination of local timezone offset to separate function