Send Xclose commands to avoid server side space leaks (#97)
The server automatically discards the result set if it was small enough to
send everything immediately, as configured with `Xreply_size`. Larger
result sets are kept on the server so the client can request the remaining
rows later using `Xexport`. It this situation, the client must explicitly destroy
the result set using `Xclose`.