shutdown the socket after sending control command
After having sent the response to a control command, the server waits
for one full second before closing the connection.  It is actually
waiting for the client to send another command. This never happens
because because the control protocol is a one-shot protocol, but the
server does it anyway, probably because the same code is also used
for normal connections.

This means the client has to wait a full second before it knows it has
received the full response.

By shutting down the write side of the socket after sending the
command we let the server know no more commands are coming so it can
shut down the connection immediately.

This speeds up the test_control.py test significantly.