Support local control connections again (#112)
* Add tests for local control connections

To check for issue #111

* Handle local control connections properly again

For some reason, monetdbd does not use the block protocol for
connections over the unix domain socket with language == 'control'.

Earlier I removed the code that dealt with this because I thought
it wasn't used anymore.  Turns out it is.

* shutdown the socket after sending control command

After having sent the response to a control command, the server waits
for one full second before closing the connection.  It is actually
waiting for the client to send another command. This never happens
because because the control protocol is a one-shot protocol, but the
server does it anyway, probably because the same code is also used
for normal connections.

This means the client has to wait a full second before it knows it has
received the full response.

By shutting down the write side of the socket after sending the
command we let the server know no more commands are coming so it can
shut down the connection immediately.

This speeds up the test_control.py test significantly.

* Silence the checkers

* Pass the passphrase to TestLocalControl

* Fix permissions of monetdbd's .merovingian socket